requirements.txt
flask
customtkinter
requests
qrcode
netifaces
screeninfo
pillow
tkinterdnd2
pyinstaller
aiortc
aiohttp
pystray
Pillow
python-socketio[client]